Tracker: Add basic detectCollision function
Detect if a collision occurs by taking into consideration the ratios
between the sizes of different objects and the thresholds defined. This
way, if two rectangles superpose, they have to have the proper scale to
be considered a collision. If they don't, they just superpose but they
belong to different planes in space.

Signed-off-by: Hernán Gonzalez <hernan.gonzalez@incluit.com>